[x86 testsuite] preserve full register across main
This test uses a call-saved register as a global variable.  It
attempts to preserve its value across main, but only the lower int
part is preserved, which is not good enough for x86_64, when the
runtime that calls main() happens to hold something in the chosen
register that is not a zero-extension from the 32-bit value, and
rightfully expects the full register to remain unchanged when main()
returns.


for  gcc/testsuite/ChangeLog

	* gcc.target/i386/20020616-1.c: Preserve full register across
	main.

git-svn-id: svn://gcc.gnu.org/svn/gcc/trunk@275329 138bc75d-0d04-0410-961f-82ee72b054a4